Moved all event handlers and related code out of control.lua and into eventhandlers.lua.
Refactored a fair amount of the code.

Changed indenting. 4 spaces, as agreed! ;)

Semantics have barely changed at all - I tested for a few minutes and everything still seems to work.
